zzz/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
python-dateutil==2.8.2
aiohttp==3.9.1
schedule==1.2.1 